Fork Update

+++ game/scripts/vscripts/relicmanager.lua
@@ -22,15 +22,21 @@ function RelicManager:Initialize()
   self.uniqueDropTable = relics.relic_type_unique
   
   for relic, weight in pairs( self.genericDropTable ) do
-	LinkLuaModifier( relic, "relics/generic/"..relic, LUA_MODIFIER_MOTION_NONE )
+	if weight ~= 0 then
+		LinkLuaModifier( relic, "relics/generic/"..relic, LUA_MODIFIER_MOTION_NONE )
+	end
   end
   
   for relic, weight in pairs( self.cursedDropTable ) do
-	LinkLuaModifier( relic, "relics/cursed/"..relic, LUA_MODIFIER_MOTION_NONE )
+	if weight ~= 0 then
+		LinkLuaModifier( relic, "relics/cursed/"..relic, LUA_MODIFIER_MOTION_NONE )
+	end
   end
   
   for relic, weight in pairs( self.uniqueDropTable ) do
-	LinkLuaModifier( relic, "relics/unique/"..relic, LUA_MODIFIER_MOTION_NONE )
+	if weight ~= 0 then
+		LinkLuaModifier( relic, "relics/unique/"..relic, LUA_MODIFIER_MOTION_NONE )
+	end
   end
   
   CustomGameEventManager:RegisterListener('player_selected_relic', Context_Wrap( RelicManager, 'ConfirmRelicSelection'))
@@ -102,14 +108,23 @@ function RelicManager:SkipRelicSelection(userid, event)
 		toNumPlayerRelics[tonumber(dropID)] = dropTable
 	end
 	
-	hero.internalRNGPools[1][toNumPlayerRelics[1]["1"]] = nil
-	hero.internalRNGPools[2][toNumPlayerRelics[1]["2"]] = nil
-	hero.internalRNGPools[3][toNumPlayerRelics[1]["3"]] = nil
-	
 	table.remove( toNumPlayerRelics, 1 )
 	relicTable[tostring(pID)] = toNumPlayerRelics
 	CustomNetTables:SetTableValue("game_info", "relic_drops", relicTable)
 	
+	for id, relic in pairs( playerRelics["1"] ) do
+		relicType = 1
+		if string.match(relic, "unique") then
+			relicType = 3
+		elseif string.match(relic, "cursed") then
+			relicType = 2
+		end
+		hero.internalRNGPools[relicType][playerRelics["1"][id]] = nil
+	end
+	
+	for i = 1, 2 do
+		hero:AddRelic( RelicManager:RollRandomGenericRelicForPlayer(pID) )
+	end
 	if hero:HasRelic("relic_unique_mysterious_hourglass") and hero:FindModifierByName("relic_unique_mysterious_hourglass"):GetStackCount() > 0 then
 		hero:FindModifierByName("relic_unique_mysterious_hourglass"):DecrementStackCount()
 		self:RollRelicsForPlayer(pID)
@@ -131,7 +146,7 @@ function RelicManager:RegisterPlayer(pID)
 	RelicManager:RollRelicsForPlayer(pID)
 end
 
-function RelicManager:RollRelicsForPlayer(pID)
+function RelicManager:RollRelicsForPlayer(pID, relicType)
 	local hero = PlayerResource:GetSelectedHeroEntity(pID)
 
 	if not hero then return end
@@ -146,9 +161,8 @@ function RelicManager:RollRelicsForPlayer(pID)
 	end
 	
 	local dropTable = {}
-	table.insert( dropTable, self:RollRandomGenericRelicForPlayer(pID) )
-	table.insert( dropTable, self:RollRandomCursedRelicForPlayer(pID) )
 	table.insert( dropTable, self:RollRandomUniqueRelicForPlayer(pID) )
+	table.insert( dropTable, self:RollRandomCursedRelicForPlayer(pID) )
 	
 	table.insert( toNumPlayerRelics, dropTable )
 
@@ -157,13 +171,15 @@ function RelicManager:RollRelicsForPlayer(pID)
 	CustomNetTables:SetTableValue("game_info", "relic_drops", relicTable)
 end
 
-function RelicManager:RollRandomGenericRelicForPlayer(pID)
+function RelicManager:RollRandomGenericRelicForPlayer(pID, notThisRelic)
 	local dropTable = {}
 	local hero = PlayerResource:GetSelectedHeroEntity(pID)
 	hero.ownedRelics = hero.ownedRelics or {}
 	for relic, weight in pairs( hero.internalRNGPools[1] ) do
-		for i = 1, weight do
-			table.insert(dropTable, relic)
+		if relic ~= notThisRelic then
+			for i = 1, weight do
+				table.insert(dropTable, relic)
+			end
 		end
 	end
 	if dropTable[1] == nil then
@@ -173,13 +189,15 @@ function RelicManager:RollRandomGenericRelicForPlayer(pID)
 	return dropTable[RandomInt(1, #dropTable)]
 end
 
-function RelicManager:RollRandomCursedRelicForPlayer(pID)
+function RelicManager:RollRandomCursedRelicForPlayer(pID, notThisRelic)
 	local dropTable = {}
 	local hero = PlayerResource:GetSelectedHeroEntity(pID)
 	hero.ownedRelics = hero.ownedRelics or {}
 	for relic, weight in pairs( hero.internalRNGPools[2] ) do
-		for i = 1, weight do
-			table.insert(dropTable, relic)
+		if relic ~= notThisRelic then
+			for i = 1, weight do
+				table.insert(dropTable, relic)
+			end
 		end
 	end
 	if dropTable[1] == nil then
@@ -189,13 +207,15 @@ function RelicManager:RollRandomCursedRelicForPlayer(pID)
 	return dropTable[RandomInt(1, #dropTable)]
 end
 
-function RelicManager:RollRandomUniqueRelicForPlayer(pID)
+function RelicManager:RollRandomUniqueRelicForPlayer(pID, notThisRelic)
 	local dropTable = {}
 	local hero = PlayerResource:GetSelectedHeroEntity(pID)
 	hero.ownedRelics = hero.ownedRelics or {}
 	for relic, weight in pairs( hero.internalRNGPools[3] ) do
-		for i = 1, weight do
-			table.insert(dropTable, relic)
+		if relic ~= notThisRelic then
+			for i = 1, weight do
+				table.insert(dropTable, relic)
+			end
 		end
 	end
 	if dropTable[1] == nil then
@@ -207,42 +227,52 @@ end
 
 function CDOTA_BaseNPC_Hero:HasRelic(relic)
 	self.ownedRelics = self.ownedRelics or {}
-	return (self.ownedRelics[relic] ~= nil)
+	return self:HasModifier(relic)
 end
 
-function RelicManager:ClearRelics(pID)
+function RelicManager:ClearRelics(pID, bHardClear)
 	local hero = PlayerResource:GetSelectedHeroEntity(pID)
 	relicCount = 0
-	for relic, item in pairs( hero.ownedRelics ) do
-		if relic ~= "relic_cursed_cursed_dice" then -- cursed dice cannot be removed
+	for item, relic in pairs( hero.ownedRelics ) do
+		if (relic == "relic_cursed_cursed_dice" and bHardClear) or relic ~= "relic_cursed_cursed_dice" then -- cursed dice cannot be removed
 			relicCount = relicCount + 1
 			hero:RemoveModifierByName( relic )
 			UTIL_Remove( EntIndexToHScript(item) )
+			hero.ownedRelics[item] = nil
 		end
 	end
 	hero.internalRNGPools[1] = self.genericDropTable
 	hero.internalRNGPools[2] = self.cursedDropTable
 	hero.internalRNGPools[3] = self.uniqueDropTable
-	hero.ownedRelics = {}
-	CustomNetTables:SetTableValue("relics", "relic_inventory_player_"..hero:entindex(), {})
+
+	CustomNetTables:SetTableValue("relics", "relic_inventory_player_"..hero:entindex(), self.ownedRelics)
 	return relicCount
 end
 
-function RelicManager:RemoveRelicOnPlayer(relic, pID)
+function RelicManager:RemoveRelicOnPlayer(relic, pID, bAll)
 	local hero = PlayerResource:GetSelectedHeroEntity(pID)
-	hero:RemoveModifierByName( relic )
-	UTIL_Remove( EntIndexToHScript(hero.ownedRelics[relic]) )
 	
-	if string.match(relic, "unique") then
-		hero.internalRNGPools[3][relic] = "1"
-	elseif string.match(relic, "cursed") then
-		hero.internalRNGPools[2][relic] = "1"
-	else
-		hero.internalRNGPools[1][relic] = "1"
+	
+	for entindex, relicName in pairs(hero.ownedRelics) do
+		if relicName == relic then
+			local item = EntIndexToHScript(entindex)
+			UTIL_Remove( item )
+			for _, modifier in ipairs( hero:FindAllModifiers() ) do
+				if modifier:GetAbility() == item then modifier:Destroy() end
+			end
+			if string.match(relicName, "unique") then
+				hero.internalRNGPools[3][relicName] = "1"
+			elseif string.match(relic, "cursed") then
+				hero.internalRNGPools[2][relicName] = "1"
+			else
+				hero.internalRNGPools[1][relicName] = "1"
+			end
+			
+			hero.ownedRelics[entindex] = nil
+			if not bAll then break end
+		end
 	end
-	
-	hero.ownedRelics[relic] = nil
-	CustomNetTables:SetTableValue("relics", "relic_inventory_player_"..hero:entindex(), hero.ownedRelics)
+	CustomNetTables:SetTableValue("relics", "relic_inventory_player_"..hero:entindex(), self.ownedRelics)
 end
 
 function CDOTA_BaseNPC_Hero:AddRelic(relic)
@@ -256,10 +286,8 @@ function CDOTA_BaseNPC_Hero:AddRelic(relic)
 		self.internalRNGPools[1][relic] = nil
 	end
 	
-	
-	
 	local relicEntity = CreateItem("item_relic_handler", nil, nil)
-	self.ownedRelics[relic] = relicEntity:entindex()
+	self.ownedRelics[relicEntity:entindex()] = relic
 	self:AddNewModifier( self, relicEntity, relic, {} )
 	
 	CustomNetTables:SetTableValue("relics", "relic_inventory_player_"..self:entindex(), self.ownedRelics)
